Add: Se agregan repositorios para seguir el patron repository Add: Se agregan clases de datos y consultas para obtener pacientes y terapeutas junto a su informacion de usuario Mod: Se cambia la implementación del id, se cambia el tipo Int por Long

+++ app/src/main/java/com/cedica/cedica/data/user/UserModel.kt
@@ -25,7 +25,7 @@ val GuestUser = User(
     ]
 )
 data class User(
-    @PrimaryKey(autoGenerate = true)  val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val role: Role,
     @ColumnInfo(collate = ColumnInfo.NOCASE) val firstName: String,
     @ColumnInfo(collate = ColumnInfo.NOCASE) val lastName: String,
@@ -41,5 +41,5 @@ interface UserDao: BaseDao<User> {
     fun getAllUsers(): Flow<List<User>>
 
     @Query("SELECT * FROM User WHERE id = :id")
-    suspend fun getByID(id: Int): User
+    suspend fun getByID(id: Long): User
 }
