+++ app/src/main/java/com/cedica/cedica/ui/home/MainMenuViewModel.kt
@@ -15,7 +15,7 @@ data class MainMenuUiState(
 )
 
 class MainMenuViewModel(
-    private val userID: Flow<Int>,
+    private val userID: Flow<Long>,
     private val db : DB
 ): ViewModel() {
